Fix support for github enterprise, no review possible
The review api is in beta and only available on github.com. As all of a
request sources methods are called once in danger local, this would try
to make a request to the review api which would fail. This pr provides a noop
review object incase the call to the review api fails with a 404.
Fixes #725