fix `--no-ansi` behavior for commands
colored2 seems to be used as drop in replacement for colored. It works
great unless you want to disable ansi codes in output as `ansi` flag is
handled by CLAide and not by Danger. Added
`Danger::Runner#adjust_colored2_output` is called from initializer
as that is only place common for commands. Otherwise some wrapper of
Runner#run will be needed (which does not seems worth it).