Use GitLab API to determine base_commit
It seems like danger's GitLab integration was build for the old API v3,
so it still needed to calculate the `base_commit` itself. With the API
v4 (which has been around since July 2016) the Merge Request actually
exposes the `base_commit`.

With this commit we use `#diff_refs#base_sha` to determine the
`base_commit` and `#diff_refs#head_sha` to determine the `head_commit`.
No need to retrieve commits anymore and calculate the SHAs from there.