Fix diff range detection logic
In many versions of GNU diff, each range can omit the comma and trailing
value s, in which case s defaults to 1. The omision has been handled
properly on the new file range, but not so on the original file range.

See https://en.wikipedia.org/wiki/Diff for more details.