Add MessageGroup class
A MessageGroup essentially represents a comment on a PR. It has a file +
line (both of which can be nil) and a list of messages (currently
Violations and Markdowns)

There's still more to do on the MessageGroup class - only #same_line?
and #<< are currently tested, and I've not written any docs for the
class yet.