Add <=> to Markdown and Violation
It compares them based on their type, then file, then line.
If both the `file`s being compared are nil, it returns 0 - comparing on
line is meaningless in this case.

The behaviour of <=> is entirely spec'd out in the
messages/violation_spec file. I will get round to Markdown's <=> spec
tomorrow