[bitbucketcloud] Require UUID to be set in environment
Bitbucket changed its API, and now accessing /users/<username> returns a
404.

We were only using that endpoint to find the UUID for identifying
whether the user that wrote a comment was danger (to decide whether to
attempt to delete it), so now we just require the UUID and don't fetch
danger's user details.

After this commit everyone using bitbucketcloud will have to provide the
DANGER_BITBUCKETCLOUD_UUID env var. It seems neater than having a
feature that presumably a lot of people are using (danger deleting its
old comments) not work unless you set that variable. But I am opening to
reworking this, if there's protest.

Fixes #1213

Also I made the specs for BitbucketCloud and BitbucketCloudAPI more
RSpec-idiomatic.