Add pr_draft? method to GitHub DSL and fix tests.
This checks the mergeable state of the PR JSON to determine whether
a PR is a draft. This doesn't add a real proper test for the method
because it didn't seem like the other methods (pr_author, pr_labels,
etc.) had dedicated tests, so I figured this didn't need one.

I can add a specific test with a fixture if we want that,
although I may need a bit of guidance on how to get the fixture to
use. :)